requirements.txt
Pillow==9.2.0
pygame==2.1.2
matplotlib==3.6.1